requirements.txt
astroid==4.0.3
blinker==1.9.0
click==8.3.1
dill==0.4.1
distlib==0.4.0
filelock==3.20.3
Flask==3.1.3
Flask-SQLAlchemy==3.1.1
gunicorn==25.1.0
isort==7.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mccabe==0.7.0
packaging==26.0
pillow==12.1.0
platformdirs==4.5.1
pylint==4.0.4
python-dotenv==1.2.1
SQLAlchemy==2.0.47
tomlkit==0.14.0
typing_extensions==4.15.0
virtualenv==20.36.1
Werkzeug==3.1.6
